+++ lib/main.dart
@@ -1,74 +1,8 @@
 import 'package:flutter/material.dart';
-import 'dart:async';
-import 'dart:convert';
-import 'package:http/http.dart' as http;
+import 'listView.dart';
 
 void main() {
   runApp(MaterialApp(
     home: StarWarsData(),
   ));
-}
-
-class StarWarsData extends StatefulWidget {
-  @override
-  StarWarsState createState() => StarWarsState();
-}
-
-class StarWarsState extends State<StarWarsData> {
-  final String url = "http://54.250.202.29/questions";
-  List data;
-
-  Future<String> getSWData() async {
-    var res =
-    await http.get(Uri.parse(url), headers: {"Accept": "application/json"});
-
-    setState(() {
-      var resBody = json.decode(res.body);
-      data = resBody;
-    });
-
-    return "Success!";
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text("Questions"),
-        backgroundColor: Colors.deepPurpleAccent,
-      ),
-      body: ListView.builder(
-        itemCount: data == null ? 0 : data.length,
-        itemBuilder: (BuildContext context, int index) {
-          return Container(
-            child: Center(
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.stretch,
-                children: <Widget>[
-                  Card(
-                    child: Container(
-                        padding: EdgeInsets.all(15.0),
-                        child: Row(
-                          children: <Widget>[
-                            Text("Question: "),
-                            Text(data[index]["question"],
-                                style: TextStyle(
-                                    fontSize: 18.0, color: Colors.black87)),
-                          ],
-                        )),
-                  ),
-                ],
-              ),
-            ),
-          );
-        },
-      ),
-    );
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    this.getSWData();
-  }
 }
